requirements.txt
addict==2.4.0
apex==0.9.10dev
basicsr==1.4.2
cityscapesscripts==2.2.3
decord==0.6.0
diffusers==0.29.2
einops==0.8.0
fairscale==0.4.13
flash_attn==2.5.9.post1
ftfy==6.2.0
horovod==0.28.1
huggingface_hub==0.23.4
iopath==0.1.10
ipython==8.12.3
jsonlines==4.0.0
lmdb==1.5.1
matplotlib==3.9.1
mc==0.0
nltk==3.8.1
numpy==2.0.0
omegaconf==2.3.0
onnx==1.16.1
opendatasets==0.1.22
packaging==24.1
pandas==2.2.2
parrots==1.0.3
peft==0.11.1
Pillow==10.4.0
prettytable==3.10.0
pycocoevalcap==1.2
pycocotools==2.0.8
python_magic==0.4.27
PyYAML==6.0.1
regex==2024.5.15
Requests==2.32.3
scikit_learn==1.5.1
scipy==1.14.0
skimage==0.0
spacy==3.7.5
tensorboardX==2.6.2.2
tensorflow==2.16.2
tifffile==2024.7.2
timm==1.0.7
torch==2.3.1
torch_tb_profiler==0.4.3
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
turbojpeg==0.0.2
visual_genome==1.1.1
webdataset==0.2.86
yapf==0.40.2
